--- front/src/containers/member/newJoin.tsx
@@ -1,5 +1,0 @@
-import * as React from "react";
-import styled from "styled-components";
-import Button from "@material-ui/core/Button";
-
-interface joinState {}
